Moved sound related code into subpackage
